#12: Fix Jpa

+++ backend/src/main/java/com/nhom4/moviereservation/model/MovieGenreId.java
@@ -2,16 +2,16 @@ package com.nhom4.moviereservation.model;
 
 import java.io.Serializable;
 
+import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import jakarta.persistence.Embeddable;
 
 @Embeddable
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class MovieGenreId implements Serializable {
-    private Long movieId;
-    private Long genreId;
+    private Integer movieId;
+    private Integer genreId;
 }
